fix: Override karpenter instance profile name
There is a pre-existing variable in `karpenter` which can override the
instance profile name, but it's only applied for the Helm chart. For
the instance profile name itself it's not used. We use
`instance_profile_name` as a default, and only if it is not set
`local.karpenter_node_iam_role_name` is instead used.